Fixed NeutralUnits never being updated, AvailableSupply overflow and tweak rock engagement distance

+++ Bot/Controller/Controller.cs
@@ -28,7 +28,8 @@ public static class Controller {
 
     public static uint CurrentSupply;
     public static uint MaxSupply;
-    public static uint AvailableSupply => MaxSupply - CurrentSupply;
+    public static int AvailableSupply => (int)(MaxSupply - CurrentSupply);
+
     public static bool IsSupplyCapped => AvailableSupply == 0;
 
     public static int AvailableMinerals;
